refactor : account service 메서드 리팩터링

+++ account/src/main/java/com/system/application/dto/BalanceResponse.java
@@ -1,6 +1,8 @@
 package com.system.application.dto;
 
+import java.math.BigDecimal;
+
 public record BalanceResponse(
-        long totalAmount
+        BigDecimal totalAmount
 ) {
 }
